Move networking code from Virtusize.kt to VirtusizeAPIService.kt
